+++ src/main/java/bootcamp/model/Message.java
