+++ src/test/java/bootcamp/model/MessageTest.java
